Add files via upload
Explored different methods for handling missing values in the development of prediction models. Used data set based on National Health and Nutrition Examination Survey runned by the Centers of Disease Control and Prevention in the USA. The main goal of the project was to predict hypertension ( high blood pressure) from a number of subject covariates and compare the input then regress classifiers to methods that handle missing values natively.